add chech for destroy callback on states

+++ engine/src/states.c
@@ -21,9 +21,12 @@ static inline void wait_vblank(void)
 void set_current_state(State *state) 
 {
     if (_current_state)
-        _current_state->destroy_callback();
-    
-        _current_state = state;
+    {
+        if(_current_state->destroy_callback) 
+            _current_state->destroy_callback();
+    }
+
+    _current_state = state;
     state->init_callback();
 }
 void states_loop()
